+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGES
 **3.0.3**
 * Support day hour field only for TimePicker
+* Support highlight first field
 
 **3.0.2**
 * Support day field for TimePicker